requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.33.0
ultralytics==8.3.72
opencv-python==4.10.0.84
mediapipe==0.10.14
paho-mqtt==2.1.0
python-multipart==0.0.20
python-dotenv==1.0.1
